feat: Mark reactions as IMAP-seen in marknoticed_chat() (#6210)
When a reaction notification is shown in the UIs, there's an option "Mark Read", but the UIs are
unaware of reactions message ids, moreover there are no `msgs` rows for reactions in the core, so
the UIs just call `marknoticed_chat()` in this case. We don't want to introduce reactions message
ids to the UIs (at least currently), but let's make reactions usual messages, just hidden, so that
the existing `\Seen` flag synchronisation mechanism works for them, and mark all reactions in the
chat as seen in marknoticed_chat().